+++ core/migrations/Migration20200917100000AuthDomainIndex.php
@@ -12,7 +12,8 @@ class Migration202000171000000AuthDomainIndex extends Base
 {
 	public function up()
 	{
-		if ($this->db->tableExists('#__auth_domain'))
+		if ($this->db->tableExists('#__auth_domain') &&
+            !$this->db->tableHasKey('#__auth_domain', 'authenticator_idx'))
 		{
 			$query = "ALTER TABLE `#__auth_domain` ADD INDEX authenticator_idx (authenticator);";
 			$this->db->setQuery($query);
@@ -22,7 +23,8 @@ class Migration202000171000000AuthDomainIndex extends Base
 
 	public function down()
 	{
-		if ($this->db->tableExists('#__auth_domain'))
+		if ($this->db->tableExists('#__auth_domain') &&
+            $this->db->tableHasKey('#__auth_domain', 'authenticator_idx'))
 		{
 			$query = "ALTER TABLE `#__auth_domain` DROP INDEX authenticator_idx;";
 			$this->db->setQuery($query);
